Remove obsolete alarm permission
