+++ app/src/main/AndroidManifest.xml
@@ -4,7 +4,6 @@
     package="com.spymag.portfoliowidget">
 
     <uses-permission android:name="android.permission.INTERNET"/>
-    <uses-permission android:name="android.permission.SCHEDULE_EXACT_ALARM"/>
 
     <application
         android:allowBackup="true"
